# backend/openapi.yaml
...
openapi: 3.1.0
info:
  title: SourceCraft Repo Health API
  version: 1.0.0
  description: 'API сервиса SourceCraft Repo Health. Файл сгенерирован из кода (app.openapi()),
    живая версия - /api/docs. Ошибки: {"error": {"code", "message"}}. Анализ асинхронный:
    202 + job_id, статус - GET /api/v1/jobs/{job_id}. scope=public - публичный анализ
    (рейтинг), scope=personal - личный по PAT пользователя.'
paths:
...
  /api/v1/repositories/{repo_id}:
    get:
      tags:
      - Analysis
      summary: Get Repository
      description: 'Отчёт по репозиторию. scope=public - публичный анализ (как в рейтинге);

        scope=personal - личный анализ по PAT пользователя (с AppSec), только для
        него.'
      operationId: get_repository_api_v1_repositories__repo_id__get
      parameters:
      - name: repo_id
        in: path
        required: true
        schema:
          type: string
          title: Repo Id
      - name: scope
        in: query
        required: false
        schema:
          type: string
          default: public
          title: Scope
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/repositories/{repo_id}/history:
    get:
      tags:
      - Analysis
      summary: Get History
      operationId: get_history_api_v1_repositories__repo_id__history_get
      parameters:
      - name: repo_id
        in: path
        required: true
        schema:
          type: string
          title: Repo Id
      - name: scope
        in: query
        required: false
        schema:
          type: string
          default: public
          title: Scope
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HistoryPoint'
                title: Response Get History Api V1 Repositories  Repo Id  History
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/repositories/analyze:
    post:
      tags:
      - Analysis
      summary: Analyze Repository
      description: 'Первичный анализ по ссылке. Идемпотентен по репозиторию: если
        анализ уже в очереди или

        выполняется - возвращается существующая задача, новая не создаётся.'
      operationId: analyze_repository_api_v1_repositories_analyze_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalyzeRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAccepted'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/repositories/{repo_id}/reanalyze:
    post:
      tags:
      - Analysis
      summary: Reanalyze Repository
      description: Повторный анализ. last_analyzed_at обновится по завершении задачи,
        а не в момент запроса.
      operationId: reanalyze_repository_api_v1_repositories__repo_id__reanalyze_post
      parameters:
      - name: repo_id
        in: path
        required: true
        schema:
          type: string
          title: Repo Id
      - name: scope
        in: query
        required: false
        schema:
          type: string
          default: public
          title: Scope
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobAccepted'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/jobs/{job_id}:
    get:
      tags:
      - Analysis
      - Jobs
      summary: Get Job
      operationId: get_job_api_v1_jobs__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/auth/yandex/callback:
    get:
      tags:
      - Auth
      summary: Yandex Callback
      operationId: yandex_callback_api_v1_auth_yandex_callback_get
      parameters:
      - name: code
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Code
      - name: state
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: State
      - name: error
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AnalyzeRequest:
      properties:
        repository_url:
          type: string
          title: Repository Url
        scope:
          type: string
          title: Scope
          default: public
      type: object
      required:
      - repository_url
      title: AnalyzeRequest
...
    Component:
      properties:
        label:
          type: string
          title: Label
        points:
          anyOf:
          - type: number
          - type: 'null'
          title: Points
        max_points:
          type: number
          title: Max Points
        value:
          type: string
          title: Value
      type: object
      required:
      - label
      - points
      - max_points
      - value
      title: Component
    Evidence:
      properties:
        type:
          type: string
          enum:
          - file
          - pipeline
          - vulnerability
          - commit
          - issue
          - merge_request
          title: Type
        ref:
          type: string
          title: Ref
        url:
          anyOf:
          - type: string
          - type: 'null'
          title: Url
      type: object
      required:
      - type
      - ref
      title: Evidence
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    HistoryPoint:
      properties:
        analyzed_at:
          type: string
          format: date-time
          title: Analyzed At
        health_score:
          anyOf:
          - type: integer
          - type: 'null'
          title: Health Score
        status:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - COMPLETED
          - PARTIAL
          - FAILED
          title: Status
      type: object
      required:
      - analyzed_at
      - health_score
      - status
      title: HistoryPoint
    Job:
      properties:
        job_id:
          type: string
          title: Job Id
        repository_id:
          type: string
          title: Repository Id
        status:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - COMPLETED
          - PARTIAL
          - FAILED
          title: Status
        progress:
          type: number
          title: Progress
          default: 0
        stage:
          anyOf:
          - type: string
          - type: 'null'
          title: Stage
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        error:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Error
        scope:
          type: string
          title: Scope
          default: public
      type: object
      required:
      - job_id
      - repository_id
      - status
      title: Job
    JobAccepted:
      properties:
        job_id:
          type: string
          title: Job Id
        repository_id:
          type: string
          title: Repository Id
        status:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - COMPLETED
          - PARTIAL
          - FAILED
          title: Status
        queued_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Queued At
        status_url:
          type: string
          title: Status Url
        scope:
          type: string
          title: Scope
          default: public
      type: object
      required:
      - job_id
      - repository_id
      - status
      - status_url
      title: JobAccepted
    JobInfo:
      properties:
        job_id:
          type: string
          title: Job Id
        status:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - COMPLETED
          - PARTIAL
          - FAILED
          title: Status
        stage:
          anyOf:
          - type: string
          - type: 'null'
          title: Stage
        progress:
          type: number
          title: Progress
          default: 0
        created_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        error:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Error
      type: object
      required:
      - job_id
      - status
      title: JobInfo
...
    Metric:
      properties:
        score:
          anyOf:
          - type: integer
          - type: 'null'
          title: Score
        status:
          type: string
          enum:
          - OK
          - NO_DATA
          title: Status
        summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary
        weight:
          type: number
          title: Weight
        base_weight:
          type: number
          title: Base Weight
        contribution:
          type: number
          title: Contribution
        components:
          items:
            $ref: '#/components/schemas/Component'
          type: array
          title: Components
          default: []
        details:
          additionalProperties: true
          type: object
          title: Details
          default: {}
        evidence:
          items:
            $ref: '#/components/schemas/Evidence'
          type: array
          title: Evidence
          default: []
      type: object
      required:
      - score
      - status
      - weight
      - base_weight
      - contribution
      title: Metric
    Metrics:
      properties:
        documentation:
          $ref: '#/components/schemas/Metric'
        ci_cd:
          $ref: '#/components/schemas/Metric'
        security:
          $ref: '#/components/schemas/Metric'
        activity:
          $ref: '#/components/schemas/Metric'
        issues:
          $ref: '#/components/schemas/Metric'
        code_health:
          $ref: '#/components/schemas/Metric'
      type: object
      required:
      - documentation
      - ci_cd
      - security
      - activity
      - issues
      - code_health
      title: Metrics
...
    PreviousAnalysis:
      properties:
        health_score:
          anyOf:
          - type: integer
          - type: 'null'
          title: Health Score
        analyzed_at:
          type: string
          format: date-time
          title: Analyzed At
      type: object
      required:
      - health_score
      - analyzed_at
      title: PreviousAnalysis
    Recommendation:
      properties:
        id:
          type: string
          title: Id
        priority:
          type: string
          enum:
          - HIGH
          - MEDIUM
          - LOW
          title: Priority
        category:
          type: string
          title: Category
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        problem:
          type: string
          title: Problem
        why_it_matters:
          type: string
          title: Why It Matters
        action:
          type: string
          title: Action
        facts:
          type: string
          title: Facts
        impact:
          anyOf:
          - type: string
          - type: 'null'
          title: Impact
        score_gain:
          type: number
          title: Score Gain
          default: 0
        evidence:
          items:
            $ref: '#/components/schemas/Evidence'
          type: array
          title: Evidence
          default: []
      type: object
      required:
      - id
      - priority
      - category
      - title
      - description
      - problem
      - why_it_matters
      - action
      - facts
      title: Recommendation
    RepositoryDetail:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        full_name:
          type: string
          title: Full Name
        url:
          type: string
          title: Url
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        language:
          anyOf:
          - type: string
          - type: 'null'
          title: Language
        likes:
          type: integer
          title: Likes
        visibility:
          type: string
          title: Visibility
        scope:
          type: string
          title: Scope
        health_score:
          anyOf:
          - type: integer
          - type: 'null'
          title: Health Score
        verdict:
          type: string
          title: Verdict
        status:
          type: string
          enum:
          - PENDING
          - IN_PROGRESS
          - COMPLETED
          - PARTIAL
          - FAILED
          title: Status
        last_analyzed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Analyzed At
        methodology_version:
          anyOf:
          - type: string
          - type: 'null'
          title: Methodology Version
        coverage:
          type: number
          title: Coverage
          default: 0
        metrics:
          anyOf:
          - $ref: '#/components/schemas/Metrics'
          - type: 'null'
        strengths:
          items:
            type: string
          type: array
          title: Strengths
          default: []
        weaknesses:
          items:
            type: string
          type: array
          title: Weaknesses
          default: []
        recommendations:
          items:
            $ref: '#/components/schemas/Recommendation'
          type: array
          title: Recommendations
          default: []
        sources:
          additionalProperties:
            type: string
          type: object
          title: Sources
          default: {}
        previous:
          anyOf:
          - $ref: '#/components/schemas/PreviousAnalysis'
          - type: 'null'
        latest_job:
          anyOf:
          - $ref: '#/components/schemas/JobInfo'
          - type: 'null'
      type: object
      required:
      - id
      - name
      - full_name
      - url
      - language
      - likes
      - visibility
      - scope
      - health_score
      - verdict
      - status
      - last_analyzed_at
      - metrics
      title: RepositoryDetail
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError